+++ include/Engine/Network/TCPClient.h
@@ -20,7 +20,7 @@ private:
     boost::asio::ip::tcp::endpoint m_Endpoint;
     boost::asio::io_service m_IOService;
     std::unique_ptr<boost::asio::ip::tcp::socket> m_Socket;
-    int readBuffer(char* data);
+    size_t readBuffer(char* data);
     PacketID m_SendPacketID = 0;
     bool m_IsConnected = false;
 };
